requirements.txt

# Transformer & model backbone
transformers>=4.57.1
timm>=1.0.22
einops>=0.8.1

# Data handling & augmentations
opencv-python>=4.12.0.88
numpy>=2.2.6
scipy>=1.16.3
pyyaml>=6.0.3
tqdm>=4.67.1

# Visualization and metrics
matplotlib>=3.10.7
open3d>=0.19.0
plyfile>=1.1.3

# Utilities
huggingface_hub>=0.36.0
pandas>=2.3.3
gdown>=5.2.0

tensorboard>=2.20.0

# # --- PyTorch Nightly Build (CUDA 12.8) ---
# # Installs nightly PyTorch, torchvision, and torchaudio with CUDA 12.8
# # --pre
# torch
# torchvision
# torchaudio
# --index-url https://download.pytorch.org/whl/nightly/cu128
